app/requirements.txt
aioredis~=2.0.1
pymongo~=4.7.2
motor~=3.4.0
requests~=2.32.2
python-dotenv~=1.0.1
celery~=5.4.0
fastapi~=0.111.0
typing_extensions~=4.11.0
pydantic~=2.7.1
pydantic_core~=2.18.2
pydantic-settings==2.3.0
Jinja2~=3.1.4
httpx~=0.27.0
python-dateutil~=2.9.0.post0
oauthlib~=3.2.2
googlemaps~=4.10.0
google-api-python-client==2.129.0
google-auth-oauthlib==1.2.0